Ref issue 47 of core
- added file functions.registry.js to store asyncSelect and formSubmitSanitization functions.
